bharosa-hyperledger/ai_service/requirements.txt
# 🧠 AI Service Dependencies (FastAPI + ML/CV)
fastapi==0.103.1
uvicorn[standard]==0.23.2
python-multipart==0.0.6
pydantic==2.3.0

# Computer Vision & ML
opencv-python==4.8.0.76
opencv-python-headless==4.8.0.76
numpy==1.24.3
pillow==10.0.0

# Face Recognition
deepface==0.0.79
tf-keras==2.15.0
tensorflow==2.15.0

# OCR
pytesseract==0.3.10

# HTTP requests
httpx==0.24.1
aiofiles==23.2.1

# Additional utilities
python-dotenv==1.0.0
